replay-scripts/requirements.txt
findspark
pyspark
azure
argparse
kafka-python
psycopg2
sqlalchemy